move environment listeners to lib/environment
...since they're mostly the same for all browsers. This makes it a bit easier to see differences between browsers, and gives us some more freedom to fix browser-specific issues without (as many) hacks.
