--- browser/index.js
@@ -1,8 +1,0 @@
-/* @flow */
-
-export {
-	sendMessage,
-	sendSynchronous,
-	addListener,
-	addInterceptor,
-} from './foreground';
