+++ firefox/manifest.json
@@ -22,12 +22,12 @@
 		}
 	},
 	"options_ui": {
-		"page": "{{../browser/options.html}}",
+		"page": "{{../lib/options.html}}",
 		"open_in_tab": true
 	},
 	"background": {
 		"scripts": [
-			"{{./background.entry.js}}"
+			"{{../lib/background.entry.js}}"
 		]
 	},
 	"content_scripts": [{
@@ -54,7 +54,7 @@
 			"https://*.reddit.com/*.json-html?*"
 		],
 		"js": [
-			"{{../lib/main.entry.js}}"
+			"{{../lib/foreground.entry.js}}"
 		],
 		"css": [
 			"{{../lib/css/res.scss}}"
