+++ lib/background.entry.js
@@ -1,0 +1,4 @@
+/* @flow */
+
+// load environment listeners
+import 'sibling-loader!./environment/background/messaging';
