+++ lib/environment/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+	"env": {
+		"webextensions": true
+	}
+}
