+++ lib/environment/background/auth.js
@@ -1,9 +1,56 @@
 /* @flow */
-/* eslint-env webextensions */
+import { apiToPromise } from '../utils/api';
+import { addListener } from './messaging';
 
-import { apiToPromise } from './api';
+addListener('authFlow', ({ domain, clientId, scope, interactive }, { index: currentIndex }) => {
+	if (process.env.BUILD_TARGET === 'chrome') {
+		// Chrome supports chrome.identity.launchAuthFlow.
+		// However--and quite inexplicably--the auth process is performed in a separate context whose cookies
+		// are cleared whenever the browser is restarted, making noninteractive auth pretty much useless.
+		// Instead, fully emulate the flow in the main context, where users will remain logged in when
+		// they expect to be, i.e., as long as they're logged into the main site.
+		// As a bonus, we can use redditenhancementsuite.com as the redirect instead of chromiumapp.org.
+		const redirectUri = 'https://redditenhancementsuite.com/oauth';
+		const url = new URL(domain);
+		url.searchParams.set('client_id', clientId);
+		url.searchParams.set('scope', scope);
+		url.searchParams.set('response_type', 'token');
+		url.searchParams.set('redirect_uri', redirectUri);
 
-export async function emulateAuthFlowInNewWindow(url: string, redirectUri: string): Promise<string> {
+		if (interactive) {
+			return emulateAuthFlowInNewWindow(url.href, redirectUri);
+		} else {
+			return emulateAuthFlowInBackground(url.href);
+		}
+	} else if (process.env.BUILD_TARGET === 'firefox') {
+		// Firefox correctly supports chrome.identity.launchAuthFlow.
+		const url = new URL(domain);
+		url.searchParams.set('client_id', clientId);
+		url.searchParams.set('scope', scope);
+		url.searchParams.set('response_type', 'token');
+		url.searchParams.set('redirect_uri', 'https://redditenhancementsuite.com/oauth');
+
+		return apiToPromise(chrome.identity.launchWebAuthFlow)({ url: url.href, interactive });
+	} else if (process.env.BUILD_TARGET === 'edge') {
+		// Edge does not support chrome.identity.launchAuthFlow.
+		// Its chrome.webRequest API does not support requests made by extensions,
+		// so we can't emulate noninteractive auth without a new tab.
+		const redirectUri = 'https://redditenhancementsuite.com/oauth';
+		const url = new URL(domain);
+		url.searchParams.set('client_id', clientId);
+		url.searchParams.set('scope', scope);
+		url.searchParams.set('response_type', 'token');
+		url.searchParams.set('redirect_uri', redirectUri);
+
+		if (interactive) {
+			return emulateAuthFlowInNewWindow(url.href, redirectUri);
+		} else {
+			return emulateAuthFlowInNewTab(url.href, redirectUri, currentIndex);
+		}
+	}
+});
+
+async function emulateAuthFlowInNewWindow(url: string, redirectUri: string): Promise<string> {
 	// Emulate interactive auth.
 	// Open a popup window, then track its progress with chrome.tabs,
 	// succeeding if it navigates to our redirect URL, and failing if it's
@@ -37,7 +84,7 @@ export async function emulateAuthFlowInNewWindow(url: string, redirectUri: strin
 	});
 }
 
-export function emulateAuthFlowInBackground(url: string): Promise<string> {
+function emulateAuthFlowInBackground(url: string): Promise<string> {
 	// Emulate noninteractive auth.
 	// Fetch the auth page. If the user is preauthorized, we will 302
 	// to the redirect URL. However, because the token is passed in the hash,
@@ -67,7 +114,7 @@ export function emulateAuthFlowInBackground(url: string): Promise<string> {
 	});
 }
 
-export async function emulateAuthFlowInNewTab(url: string, redirectUri: string, currentTabIndex: number): Promise<string> {
+async function emulateAuthFlowInNewTab(url: string, redirectUri: string, currentTabIndex: number): Promise<string> {
 	// Emulate noninteractive auth.
 	// Open a background tab and track its progress with chrome.tabs,
 	// succeeding if it navigates to the redirect URL immediately.
