+++ lib/environment/background/i18n.js
@@ -1,0 +1,6 @@
+/* @flow */
+
+import { getLocaleDictionary } from '../../../locales';
+import { addListener } from './messaging';
+
+addListener('i18n', locale => getLocaleDictionary(locale));
