+++ lib/environment/background/session.js
@@ -1,0 +1,22 @@
+/* @flow */
+
+import { addListener } from './messaging';
+
+const session = new Map();
+addListener('session', ([operation, key, value]) => {
+	switch (operation) {
+		case 'get':
+			return session.get(key);
+		case 'set':
+			session.set(key, value);
+			break;
+		case 'delete':
+			return session.delete(key);
+		case 'has':
+			return session.has(key);
+		case 'clear':
+			return session.clear();
+		default:
+			throw new Error(`Invalid session operation: ${operation}`);
+	}
+});
