+++ lib/environment/foreground/ajax.js
@@ -1,8 +1,8 @@
 /* @flow */
 
 import _ from 'lodash';
-import { sendMessage } from '../../browser';
-import { loggedInUserHash } from '../utils';
+import { loggedInUserHash } from '../../utils';
+import { sendMessage } from './messaging';
 import * as XhrCache from './xhrCache';
 
 type ResponseType = 'text' | 'json' | 'raw';
