+++ lib/environment/foreground/auth.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
+import { sendMessage } from './messaging';
 import * as Permissions from './permissions';
 
 export async function launchAuthFlow(
