+++ lib/environment/foreground/download.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
+import { sendMessage } from './messaging';
 
 export function download(url: string, filename?: string) {
 	return sendMessage('download', {
