+++ lib/environment/foreground/history.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
+import { sendMessage } from './messaging';
 import { isPrivateBrowsing } from './privateBrowsing';
 
 export async function addURLToHistory(url: string): Promise<void> {
