+++ lib/environment/foreground/i18n.js
@@ -1,9 +1,9 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
-import { CACHED_LANG_KEY, CACHED_MESSAGES_KEY, CACHED_MESSAGES_TOKEN_KEY } from '../constants/localStorage';
-import { rawLocale } from '../utils/localization';
-import buildToken from 'exec-loader!../../build/buildToken';
+import buildToken from 'exec-loader!../../../build/buildToken';
+import { CACHED_LANG_KEY, CACHED_MESSAGES_KEY, CACHED_MESSAGES_TOKEN_KEY } from '../../constants/localStorage';
+import { rawLocale } from '../../utils/localization';
+import { sendMessage } from './messaging';
 
 let messages;
