+++ lib/environment/foreground/id.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+export function getExtensionId(): string {
+	return chrome.runtime.id;
+}
