+++ lib/environment/foreground/pageAction.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { addListener, sendMessage } from '../../browser';
+import { addListener, sendMessage } from './messaging';
 
 const clickListeners = [];
