+++ lib/environment/foreground/permissions.js
@@ -1,8 +1,8 @@
 /* @flow */
 
 import _ from 'lodash';
-import { sendMessage } from '../../browser';
-import { mutex } from '../utils';
+import { mutex } from '../../utils';
+import { sendMessage } from './messaging';
 
 function filterPerms(perms) {
 	const permissions = perms.filter(p => !p.includes('://') && p !== '<all_urls>');
