+++ lib/environment/foreground/privateBrowsing.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+export function isPrivateBrowsing(): boolean {
+	return chrome.extension.inIncognitoContext;
+}
