+++ lib/environment/foreground/session.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
+import { sendMessage } from './messaging';
 
 export function get(key: string): Promise<any | void> {
 	return sendMessage('session', ['get', key]);
@@ -10,10 +10,10 @@ export function set(key: string, value: mixed) {
 	return sendMessage('session', ['set', key, value]);
 }
 
-export function _delete(key: string) {
+export function delete_(key: string) {
 	return sendMessage('session', ['delete', key]);
 }
-export { _delete as delete };
+export { delete_ as delete };
 
 export function has(key: string): Promise<boolean> {
 	return sendMessage('session', ['has', key]);
