+++ lib/environment/foreground/storage.js
@@ -1,26 +1,43 @@
 /* @flow */
 
 import _ from 'lodash';
-import { sendMessage, sendSynchronous } from '../../browser';
+import { extendDeep, keyedMutex } from '../../utils';
+import { apiToPromise } from '../utils/api';
+import { sendMessage } from './messaging';
+
+const __set = apiToPromise((items, callback) => chrome.storage.local.set(items, callback));
+const _set = (key, value) => __set({ [key]: value });
+const __get = apiToPromise((keys, callback) => chrome.storage.local.get(keys, callback));
+const _get = async (key, defaultValue = null) => (await __get({ [key]: defaultValue }))[key];
+const _delete = apiToPromise((keys, callback) => chrome.storage.local.remove(keys, callback));
+const _clear = apiToPromise(callback => chrome.storage.local.clear(callback));
+
+const withLockOn = keyedMutex(<T>(key, fn: () => T): T => fn());
 
 export function get(key: string): Promise<any | null> {
-	return sendSynchronous('storage', ['get', key]);
+	return withLockOn(key, () => _get(key, null));
 }
 
 export function getAll(): Promise<{ [string]: mixed }> {
-	return sendSynchronous('storage', ['getAll']);
+	return withLockOn(null, () => __get(null));
 }
 
 export function getMultiple<T: string>(keys: T[]): Promise<{ [T]: any | null }> {
-	return sendSynchronous('storage', ['getMultiple', keys]);
+	return withLockOn(null, () => {
+		const defaults = {};
+		for (const k of keys) {
+			defaults[k] = null;
+		}
+		return __get(defaults);
+	});
 }
 
 export function set(key: string, value: mixed): Promise<void> {
-	return sendSynchronous('storage', ['set', key, value]);
+	return withLockOn(key, () => _set(key, value));
 }
 
 export function setMultiple(valueMap: { +[string]: mixed }): Promise<void> {
-	return sendSynchronous('storage', ['setMultiple', null, valueMap]);
+	return withLockOn(null, () => __set(valueMap));
 }
 
 function compareAndSet(key: string, oldValue: mixed, newValue: mixed): Promise<boolean> {
@@ -31,14 +48,20 @@ function compareAndSet(key: string, oldValue: mixed, newValue: mixed): Promise<b
  * Deeply patches an object in storage, like extendDeep().
  */
 function patch(key: string, value: { [string]: mixed }): Promise<void> {
-	return sendSynchronous('storage', ['patch', key, value]);
+	return withLockOn(key, async () => {
+		const extended = extendDeep(await _get(key) || {}, value);
+		return _set(key, extended);
+	});
 }
 
 /*
  * Shallowly patches an object in storage, like Object.assign().
  */
 function patchShallow(key: string, value: { [string]: mixed }): Promise<void> {
-	return sendSynchronous('storage', ['patchShallow', key, value]);
+	return withLockOn(key, async () => {
+		const extended = Object.assign(await _get(key) || {}, value);
+		return _set(key, extended);
+	});
 }
 
 /*
@@ -47,24 +70,37 @@ function patchShallow(key: string, value: { [string]: mixed }): Promise<void> {
  * will `delete userTaggerStoredValue.username.tag`
  */
 function deletePaths(key: string, paths: string[][]): Promise<void> {
-	return sendSynchronous('storage', ['deletePaths', key, paths]);
+	return withLockOn(key, async () => {
+		const stored = await _get(key);
+		for (const path of paths) {
+			path.reduce((obj, key, i, { length }) => {
+				if (!obj) return;
+				if (i < length - 1) return obj[key];
+				delete obj[key];
+			}, stored);
+		}
+		return _set(key, stored);
+	});
 }
 
-function _delete(keys: string | string[]): Promise<void> {
-	return sendSynchronous('storage', ['delete', keys]);
+function delete_(keys: string | string[]): Promise<void> {
+	return withLockOn(typeof keys === 'string' ? keys : null, () => _delete(keys));
 }
-export { _delete as delete };
+export { delete_ as delete };
 
 export function has(key: string): Promise<boolean> {
-	return sendSynchronous('storage', ['has', key]);
+	return withLockOn(key, async () => {
+		const sentinel = Math.random();
+		return (await _get(key, sentinel)) !== sentinel;
+	});
 }
 
 export function keys(): Promise<string[]> {
-	return sendSynchronous('storage', ['keys']);
+	return withLockOn(null, async () => Object.keys(await __get(null)));
 }
 
 export function clear(): Promise<void> {
-	return sendSynchronous('storage', ['clear']);
+	return withLockOn(null, () => _clear());
 }
 
 class Wrapper<T> {
@@ -97,7 +133,7 @@ class Wrapper<T> {
 	}
 
 	delete(): Promise<void> {
-		return _delete(this._key());
+		return delete_(this._key());
 	}
 
 	has(): Promise<boolean> {
@@ -170,7 +206,7 @@ class PrefixWrapper<T> {
 	}
 
 	delete(keys: string | string[]): Promise<void> {
-		return _delete([].concat(keys).map(k => this._keyGen(k)));
+		return delete_([].concat(keys).map(k => this._keyGen(k)));
 	}
 
 	has(key: string): Promise<boolean> {
@@ -242,7 +278,7 @@ class BlobWrapper<T> {
 	}
 
 	clear(): Promise<void> {
-		return _delete(this._rootKey);
+		return delete_(this._rootKey);
 	}
 }
