+++ lib/environment/foreground/tabs.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
+import { sendMessage } from './messaging';
 
 export function openNewTab(url: string, focus?: boolean = true) {
 	return openNewTabs(focus, url);
