+++ lib/environment/foreground/xhrCache.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from '../../browser';
+import { sendMessage } from './messaging';
 
 export function set(key: string, value: mixed) {
 	return sendMessage('XHRCache', ['set', key, value]);
@@ -10,10 +10,10 @@ export function check(key: string, maxAge?: number /* milliseconds */): Promise<
 	return sendMessage('XHRCache', ['check', key, maxAge]);
 }
 
-function _delete(key: string) {
+function delete_(key: string) {
 	return sendMessage('XHRCache', ['delete', key]);
 }
-export { _delete as delete };
+export { delete_ as delete };
 
 export function clear() {
 	return sendMessage('XHRCache', ['clear']);
