--- lib/environment/id.js
@@ -1,7 +1,0 @@
-/* @flow */
-
-import { sendSynchronous } from '../../browser';
-
-export function getExtensionId(): string {
-	return sendSynchronous('extensionId');
-}
