+++ lib/environment/index.js
@@ -1,16 +1,16 @@
 /* @flow */
 
-export { addURLToHistory, isURLVisited } from './history';
-export { ajax } from './ajax';
-export { download } from './download';
-export { getExtensionId } from './id';
-export { i18n, _loadI18n } from './i18n';
-export { isPrivateBrowsing } from './privateBrowsing';
-export { launchAuthFlow } from './auth';
-export { multicast } from './multicast';
-export { openNewTab, openNewTabs } from './tabs';
-export * as PageAction from './pageAction';
-export * as Permissions from './permissions';
-export * as Session from './session';
-export * as Storage from './storage';
-export * as XhrCache from './xhrCache';
+export { addURLToHistory, isURLVisited } from './foreground/history';
+export { ajax } from './foreground/ajax';
+export { download } from './foreground/download';
+export { getExtensionId } from './foreground/id';
+export { i18n, _loadI18n } from './foreground/i18n';
+export { isPrivateBrowsing } from './foreground/privateBrowsing';
+export { launchAuthFlow } from './foreground/auth';
+export { multicast } from './foreground/multicast';
+export { openNewTab, openNewTabs } from './foreground/tabs';
+export * as PageAction from './foreground/pageAction';
+export * as Permissions from './foreground/permissions';
+export * as Session from './foreground/session';
+export * as Storage from './foreground/storage';
+export * as XhrCache from './foreground/xhrCache';
