--- lib/environment/privateBrowsing.js
@@ -1,7 +1,0 @@
-/* @flow */
-
-import { sendSynchronous } from '../../browser';
-
-export function isPrivateBrowsing(): boolean {
-	return sendSynchronous('isPrivateBrowsing');
-}
