+++ lib/environment/utils/api.js
@@ -1,7 +1,5 @@
 /* @flow */
 
-/* eslint-env webextensions */
-
 export function apiToPromise(func: (...args: mixed[]) => void): (...args: mixed[]) => Promise<any> {
 	return (...args) =>
 		new Promise((resolve, reject) =>
