--- lib/main.entry.js
@@ -1,5 +1,0 @@
-/* @flow */
-
-import { init } from './core/init';
-
-init();
