+++ lib/modules/pageNavigator.js
@@ -4,7 +4,7 @@ import _ from 'lodash';
 import { $ } from '../vendor';
 import { Module } from '../core/module';
 import { getHeaderOffset, isPageType, scrollTo, Thing, frameThrottle, string } from '../utils';
-import { i18n } from '../environment/i18n';
+import { i18n } from '../environment';
 import * as Floater from './floater';
 import * as SettingsNavigation from './settingsNavigation';
 import * as SelectedEntry from './selectedEntry';
