+++ package.json
@@ -127,6 +127,7 @@
     "sass-lint": "1.12.1",
     "sass-loader": "6.0.6",
     "semver": "5.4.1",
+    "sibling-loader": "1.2.0",
     "spawn-loader": "4.0.0",
     "url-loader": "0.6.2",
     "webpack": "3.10.0",
